refactor: move setMoneyRequestReceipt, navigateToStartStepIfScanFileCannotBeRead, checkIfScanFileCanBeRead to IOU/Receipt.ts
Move additional receipt-related functions to the Receipt module:
- setMoneyRequestReceipt: sets receipt data on transaction drafts
- navigateToStartStepIfScanFileCannotBeRead: scan file validation with navigation
- checkIfScanFileCanBeRead: scan file readability check

Update 11 consumer files and 1 test mock to import from the new path.
