+++ src/pages/iou/request/step/IOURequestStepScan/index.native.tsx
@@ -36,8 +36,8 @@ import StepScreenWrapper from '@pages/iou/request/step/StepScreenWrapper';
 import withFullTransactionOrNotFound from '@pages/iou/request/step/withFullTransactionOrNotFound';
 import withWritableReportOrNotFound from '@pages/iou/request/step/withWritableReportOrNotFound';
 import variables from '@styles/variables';
-import {setMoneyRequestReceipt, updateLastLocationPermissionPrompt} from '@userActions/IOU';
-import {replaceReceipt} from '@userActions/IOU/Receipt';
+import {updateLastLocationPermissionPrompt} from '@userActions/IOU';
+import {replaceReceipt, setMoneyRequestReceipt} from '@userActions/IOU/Receipt';
 import {buildOptimisticTransactionAndCreateDraft} from '@userActions/TransactionEdit';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
