+++ src/pages/media/AttachmentModalScreen/routes/TransactionReceiptModalContent.tsx
@@ -12,8 +12,8 @@ import useNetwork from '@hooks/useNetwork';
 import useOnyx from '@hooks/useOnyx';
 import usePolicy from '@hooks/usePolicy';
 import useThemeStyles from '@hooks/useThemeStyles';
-import {navigateToStartStepIfScanFileCannotBeRead, removeMoneyRequestOdometerImage, setMoneyRequestOdometerImage, setMoneyRequestReceipt} from '@libs/actions/IOU';
-import {detachReceipt, replaceReceipt} from '@libs/actions/IOU/Receipt';
+import {removeMoneyRequestOdometerImage, setMoneyRequestOdometerImage} from '@libs/actions/IOU';
+import {detachReceipt, navigateToStartStepIfScanFileCannotBeRead, replaceReceipt, setMoneyRequestReceipt} from '@libs/actions/IOU/Receipt';
 import {openReport} from '@libs/actions/Report';
 import cropOrRotateImage from '@libs/cropOrRotateImage';
 import fetchImage from '@libs/fetchImage';
